+++ internal/introspector.go
@@ -31,10 +31,11 @@ func (_m *MockTokenIntrospector) EXPECT() *_MockTokenIntrospectorRecorder {
 	return _m.recorder
 }
 
-func (_m *MockTokenIntrospector) IntrospectToken(_param0 context.Context, _param1 string, _param2 fosite.TokenType, _param3 fosite.AccessRequester, _param4 []string) error {
+func (_m *MockTokenIntrospector) IntrospectToken(_param0 context.Context, _param1 string, _param2 fosite.TokenType, _param3 fosite.AccessRequester, _param4 []string) (fosite.TokenType, error) {
 	ret := _m.ctrl.Call(_m, "IntrospectToken", _param0, _param1, _param2, _param3, _param4)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(fosite.TokenType)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 func (_mr *_MockTokenIntrospectorRecorder) IntrospectToken(arg0, arg1, arg2, arg3, arg4 interface{}) *gomock.Call {